Table view styling, plus some tweaks to the layout for mobile
